# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SponsorshipOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /performance/verify/{milestoneId}:
    post:
      operationId: verifyMilestoneCompliance
      tags: [performance]
      summary: Run YouTube compliance check for a milestone
      parameters:
        - name: milestoneId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Compliance check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceResult"
  /performance/report/{contractId}:
    post:
      operationId: generatePerformanceReport
      tags: [performance]
      summary: Generate AI performance report for a contract
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceReport"
    get:
      operationId: getPerformanceReport
      tags: [performance]
      summary: Get saved performance report
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Performance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceReport"
  /shared/report/{token}:
    get:
      operationId: getPublicReport
      tags: [performance]
      summary: Get public performance report by share token (no auth)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public performance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceReport"

  # SENTIMENT
  /sentiment/analyze:
    post:
      operationId: analyzeSentiment
      tags: [sentiment]
      summary: Analyze brand communication sentiment with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SentimentAnalyzeBody"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunicationAnalysis"
  /sentiment/{contractId}:
    get:
      operationId: listSentimentAnalyses
      tags: [sentiment]
      summary: Get sentiment analysis history for a contract
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunicationAnalysis"

  # YOUTUBE
...
  /notion/export/contract/{contractId}:
    post:
      operationId: exportContractToNotion
      tags: [notion]
      summary: Export contract to Notion
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotionExportBody"
      responses:
        "200":
          description: Export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotionExportResult"
  /notion/export/milestones/{contractId}:
    post:
      operationId: exportMilestonesToNotion
      tags: [notion]
      summary: Export milestones to Notion
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotionExportBody"
      responses:
        "200":
          description: Export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotionExportResult"
  /notion/export/report/{contractId}:
    post:
      operationId: exportReportToNotion
      tags: [notion]
      summary: Export performance report to Notion
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotionExportBody"
      responses:
        "200":
          description: Export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotionExportResult"

  # SETTINGS
...
components:
  schemas:
...
    ComplianceResult:
      type: object
      properties:
        milestoneId:
          type: string
        complianceStatus:
          type: string
        contentUrl:
          type: string
        matchedVideo:
          type: object
        checks:
          type: object
      required: [milestoneId, complianceStatus]
    PerformanceReport:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        campaignName:
          type: string
        brandName:
          type: string
        reportData:
          type: object
        complianceScore:
          type: integer
        sharedToken:
          type: string
        generatedAt:
          type: string
      required: [id, contractId, complianceScore, sharedToken, generatedAt]
    SentimentAnalyzeBody:
      type: object
      properties:
        contractId:
          type: string
        rawInput:
          type: string
      required: [rawInput]
    CommunicationAnalysis:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        rawInput:
          type: string
        sentiment:
          type: string
          enum: [positive, neutral, negative, mixed]
        urgencyLevel:
          type: string
          enum: [low, medium, high, critical]
        paymentRisk:
          type: string
          enum: [low, medium, high]
        scopeCreepRisk:
          type: string
          enum: [low, medium, high]
        toneRecommendation:
          type: string
        suggestedReply:
          type: string
        fullAnalysis:
          type: object
        createdAt:
          type: string
      required: [id, rawInput, sentiment, urgencyLevel, paymentRisk, scopeCreepRisk, createdAt]
...
    NotionExportBody:
      type: object
      properties:
        databaseId:
          type: string
      required: [databaseId]
    NotionExportResult:
      type: object
      properties:
        notionPageId:
          type: string
        notionUrl:
          type: string
        status:
          type: string
      required: [status]